app/requirements.txt
aiofiles==0.8.0
altair==4.2.0
anyio==3.6.1
asgiref==3.5.2
asttokens==2.0.5
attrs==21.4.0
backcall==0.2.0
backports.zoneinfo==0.2.1
catalogue==2.0.7
charset-normalizer==2.1.0
click==8.1.3
colorama==0.4.5
commonmark==0.9.1
cycler==0.11.0
decorator==5.1.1
dill==0.3.5.1
Django==4.0.6
efficientnet-pytorch==0.7.1
entrypoints==0.4
executing==0.8.3
fastapi==0.78.0
fastcore==1.5.4
fonttools==4.34.4
h11==0.13.0
idna==3.3
imageio==2.19.3
imgaug==0.4.0
importlib-resources==5.8.0
ipython==8.4.0
jedi==0.18.1
Jinja2==3.1.2
jsonlines==3.1.0
jsonschema==4.7.2
kiwisolver==1.4.3
loguru==0.6.0
MarkupSafe==2.1.1
matplotlib==3.5.2
matplotlib-inline==0.1.3
networkx==2.8.4
numpy==1.23.1
opencv-python==4.6.0.66
packaging==21.3
pandas==1.4.3
parso==0.8.3
pickleshare==0.7.5
Pillow==9.2.0
prompt-toolkit==3.0.30
pure-eval==0.2.2
pydantic==1.9.1
Pygments==2.12.0
pyparsing==3.0.9
pyrsistent==0.18.1
python-dateutil==2.8.2
python-multipart==0.0.5
pytz==2022.1
PyWavelets==1.3.0
PyYAML==6.0
requests==2.28.1
rich==12.5.1
scikit-image==0.19.3
scipy==1.8.1
Shapely==1.8.2
six==1.16.0
sniffio==1.2.0
sqlparse==0.4.2
srsly==2.4.3
stack-data==0.3.0
starlette==0.19.1
tifffile==2022.5.4
toolz==0.12.0
torch==1.12.0
torch-snippets==0.496
torch-summary==1.4.5
torchvision==0.13.0
tqdm==4.64.0
traitlets==5.3.0
typing==3.7.4.3
typing_extensions==4.3.0
tzdata==2022.1
urllib3==1.26.10
uvicorn==0.18.2
wasabi==0.9.1
wcwidth==0.2.5
win32-setctime==1.1.0
wincertstore==0.2
zipp==3.8.1